+++ tests/integration/test_keeper_memory_soft_limit_ratio/configs/keeper_config1_reload.xml
@@ -11,9 +11,6 @@
 
     <keeper_server>
         <tcp_port>2181</tcp_port>
-        <availability_zone>
-            <value>az-zoo1</value>
-        </availability_zone>
         <server_id>1</server_id>
         <max_memory_usage_soft_limit_ratio>0.9</max_memory_usage_soft_limit_ratio>
 
@@ -24,7 +21,6 @@
             <force_sync>false</force_sync>
             <election_timeout_lower_bound_ms>2000</election_timeout_lower_bound_ms>
             <election_timeout_upper_bound_ms>4000</election_timeout_upper_bound_ms>
-
             <async_replication>1</async_replication>
         </coordination_settings>
